+++ app/src/main/java/com/dan/school/DataViewModel.kt
@@ -3,6 +3,8 @@ package com.dan.school
 import android.app.Application
 import androidx.lifecycle.*
 import androidx.sqlite.db.SimpleSQLiteQuery
+import com.dan.school.data.ItemDatabase
+import com.dan.school.data.ItemRepository
 import com.dan.school.models.DateItem
 import com.dan.school.models.Item
 import kotlinx.coroutines.Dispatchers
