+++ app/src/main/java/com/dan/school/fragments/AgendaTomorrowFragment.kt
@@ -15,6 +15,7 @@ import com.dan.school.School.categoryUncheckedIcons
 import com.dan.school.adapters.BaseItemListAdapter
 import com.dan.school.adapters.ItemListAdapter
 import com.dan.school.databinding.FragmentAgendaTomorrowBinding
+import com.dan.school.interfaces.ItemClickListener
 import com.dan.school.models.Item
 import com.dan.school.models.Subtask
 import java.text.SimpleDateFormat
