+++ app/src/main/java/com/dan/school/fragments/HomeFragment.kt
@@ -11,6 +11,7 @@ import androidx.fragment.app.Fragment
 import com.dan.school.*
 import com.dan.school.adapters.HomeworkExamTaskTabLayoutAdapter
 import com.dan.school.databinding.FragmentHomeBinding
+import com.dan.school.interfaces.ItemClickListener
 import com.dan.school.models.Item
 import com.google.android.material.tabs.TabLayout
