+++ app/src/main/java/com/dan/school/fragments/OverviewFragment.kt
@@ -8,7 +8,7 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.dan.school.ItemClickListener
+import com.dan.school.interfaces.ItemClickListener
 import com.dan.school.MainActivity
 import com.dan.school.R
 import com.dan.school.School
