+++ app/src/main/java/com/dan/school/settings/BackupFragment.kt
@@ -20,7 +20,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.dan.school.*
 import com.dan.school.adapters.BackupListAdapter
 import com.dan.school.authentication.AuthenticationActivity
+import com.dan.school.data.ItemDatabase
 import com.dan.school.databinding.FragmentBackupBinding
+import com.dan.school.interfaces.BackupItemClickListener
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.textfield.TextInputLayout
 import com.google.firebase.auth.FirebaseAuth
